+++ mobile/cmp/menu/impl/Menu.ts
@@ -5,13 +5,13 @@
  * Copyright © 2025 Extremely Heavy Industries Inc.
  */
 import {div, hspacer, vbox} from '@xh/hoist/cmp/layout';
-import {hoistCmp, HoistModel, useLocalModel, MenuItem, MenuItemLike} from '@xh/hoist/core';
+import {hoistCmp, HoistModel, useLocalModel, MenuItemLike, isMenuItem} from '@xh/hoist/core';
 import {listItem} from '@xh/hoist/kit/onsen';
 import {makeObservable, bindable} from '@xh/hoist/mobx';
 import {filterConsecutiveMenuSeparators, isOmitted} from '@xh/hoist/utils/impl';
 import classNames from 'classnames';
-import {clone, isEmpty, isString} from 'lodash';
-import {isValidElement, ReactNode, useEffect} from 'react';
+import {clone, isEmpty} from 'lodash';
+import {ReactNode, useEffect} from 'react';
 
 import './Menu.scss';
 
@@ -106,8 +106,4 @@ class LocalMenuModel extends HoistModel {
                 });
             });
     }
-}
-
-function isMenuItem(item: MenuItemLike): item is MenuItem {
-    return !isString(item) && !isValidElement(item);
 }